Replace utils.py with semantically revelant file names
