+++ pisa_utils/parsers.py
@@ -9,6 +9,7 @@ import pandas as pd
 from pydantic_core import ValidationError
 import xmltodict
 
+from pisa_utils.file_io import open_compressed, save_json
 from pisa_utils.models.data_models import (
     Complex,
     ComplexExtended,
@@ -19,12 +20,10 @@ from pisa_utils.models.data_models import (
 )
 from pisa_utils.models.models import AllowedModelFileFormats
 from pisa_utils.models.validation_error_handlers import trigger_helpful_validation_error
-from pisa_utils.utils import (
+from pisa_utils.field_handlers import (
     extract_ligand_contents,
     id_is_ligand,
     is_int_or_float,
-    open_compressed,
-    save_json,
 )
 
 LOGGER = logging.getLogger(__name__)
